backend/requirements.txt
Django==3.2.3
djangorestframework==3.14.0
django-filter==23.2
djoser==2.2.0
gunicorn==20.1.0
openpyxl==3.1.2
Pillow==9.5.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0
sorl-thumbnail==12.9.0
webcolors==1.11.1